drm: vkms: Refactor the plane composer to accept new formats
Currently the blend function only accepts XRGB_8888 and ARGB_8888
as a color input.

This patch refactors all the functions related to the plane composition
to overcome this limitation.

Now the blend function receives a struct `vkms_pixel_composition_functions`
containing two handlers.

One will generate a buffer of each line of the frame with the pixels
converted to ARGB16161616. And the other will take this line buffer,
do some computation on it, and store the pixels in the destination.

The first handler behavior has been totally described previously.

The second is more interesting, as it has to perform two roles
depending on where it is called in the code.

The first is to convert(if necessary) the data received in the
`line_buffer` and write in the memory using the `src_composer`
and `dst_composer`.

The second role is to perform the `alpha_blend`. So, it takes the
pixels using `line_buffer` and `dst_composer`, executes the blend,
and stores the result back using the `dst_composer` and `src_composer`.

The per-line implementation was chosen for performance reasons.
The per-pixel functions were having performance issues due to indirect
function call overhead.

The per-line code trades off memory for execution time. The `line_buffer`
allows us to diminish the number of function calls.

Results in the IGT test `kms_cursor_crc`:

|                     Frametime                       |
|:---------------:|:---------:|:----------:|:--------:|
|  implmentation  |  Current  |  Per-pixel | Per-line |
| frametime range |  8~22 ms  |  32~56 ms  |  6~20 ms |
|     Average     |  10.0 ms  |   35.8 ms  |  9.0 ms  |

Reported-by: kernel test robot <lkp@intel.com>
Signed-off-by: Igor Torrente <igormtorrente@gmail.com>